+++ processor/src/test/resources/input/com/example/repository/RepositoryWithDetachNone.java
@@ -1,0 +1,31 @@
+package com.example.repository;
+
+import arez.annotations.ArezComponent;
+import arez.annotations.Observable;
+import arez.annotations.Repository;
+import javax.annotation.Nonnull;
+
+@Repository( detach = Repository.DetachType.NONE )
+@ArezComponent
+public abstract class RepositoryWithDetachNone
+{
+  @Nonnull
+  private String _name;
+
+  RepositoryWithDetachNone( @Nonnull final String name )
+  {
+    _name = name;
+  }
+
+  @Observable
+  @Nonnull
+  public String getName()
+  {
+    return _name;
+  }
+
+  public void setName( @Nonnull final String name )
+  {
+    _name = name;
+  }
+}